\name{deviance.gl1ce}
\alias{deviance.gl1ce}
\title{Deviance Method for `gl1ce' Objects}
\usage{
deviance.gl1ce(object, \dots)
}
\description{
  This is a method of the generic function \code{\link{deviance}()} for
  objects inheriting from class \code{gl1ce} (see
  \code{help(\link{gl1ce.object})}).
}
\arguments{
  \item{object}{an object inheriting from class \code{gl1ce}.}
  \item{\dots}{possibly further arguments (none at the moment).}
}
\seealso{\code{\link{deviance}} for the general
  behavior of this function and for the interpretation of \code{object}.
}
\keyword{models}
% Converted by Sd2Rd version 1.21.
