\name{deviance.l1ce}
\alias{deviance.l1ce}
\alias{deviance.l1celist}
\title{Deviance Method for `l1ce' and `l1celist' Objects}
\usage{
deviance.l1ce(object, \dots)
deviance.l1celist(object, \dots)
}
\arguments{
  \item{object}{an object inheriting from class \code{l1ce} or
    \code{l1celist}, respectively.}
  \item{\dots}{possibly further arguments (none at the moment).}
}
\description{
  These are methods of the generic function \code{\link{deviance}()} for
  objects inheriting from class \code{l1ce} or \code{l1celist} (see
  \code{help(\link{l1ce.object})} and \code{help(\link{l1celist.object})}).
}
\seealso{\code{\link{deviance}} for the general
  behavior of this function and for the interpretation of \code{object}.
}
\keyword{models}
% Converted by Sd2Rd version 1.21.
