\name{fitted.l1ce}
\alias{fitted.l1ce}
\alias{fitted.l1celist}
\title{Fitted Values for `l1ce', `l1celist' and `gl1ce' Objects}
\description{
  These are methods of the generic function \code{\link{fitted}()} for
  objects inheriting from class \code{l1ce} or \code{l1celist} (see
  \code{help(\link{l1ce.object})} and \code{help(\link{l1celist.object})}).
}
\usage{
fitted.l1ce(object, \dots)
fitted.l1celist(object, \dots)
}
\arguments{
  \item{object}{an object inheriting from class \code{l1ce} or
    \code{l1celist}, e.g., an \code{gl1ce} one (see
    \code{\link{gl1ce.object}}).}
  \item{\dots}{further potential arguments passed to methods.}
}
\seealso{\code{\link{fitted}} for the general
  behavior of this function and for the interpretation of \code{object}.
}
\keyword{models}
