% Copyright (C) 1998
% Berwin A. Turlach <bturlach@stats.adelaide.edu.au>
% Bill Venables <wvenable@stats.adelaide.edu.au>
% $Id: is.formula.d,v 1.2 1999/12/01 05:12:09 bturlach Exp $
% --> ../COPYRIGHT for more details
\name{is.formula}
\alias{is.formula}
\title{Tests for Formula Objects}
\usage{
is.formula(x)
}
\arguments{
  \item{x}{an \R object to be tested.}
}
\description{
  \code{is.formula} returns \code{TRUE} if \code{x} is an object of
  class \code{"formula"}, and \code{FALSE} otherwise.
}
\keyword{utilities}
