\name{lasso-internal}
\alias{opt.covmat}
\alias{tib.covmat}
\title{Internal lasso functions}
\description{
  Internal lasso functions.
}
\usage{
opt.covmat(xtx, xtr, bound, Lagrangian)
tib.covmat(xtx, gen.inverse.diag, beta, Lagrangian)
}
\details{
  These are not (yet?) to be called by the user.
}
\keyword{internal}
