% Copyright (C) 1998
% Berwin A. Turlach <bturlach@stats.adelaide.edu.au>
% Bill Venables <wvenable@stats.adelaide.edu.au>
% $Id: merge.formula.d,v 1.2 1999/12/01 05:11:12 bturlach Exp $
% --> ../COPYRIGHT for more details
\name{merge.formula}
\alias{merge.formula}
\title{Merge Formula With Right Hand Side of Second Formula}
\usage{
merge.formula(x, y, \dots)
}
\arguments{
  \item{x,y}{formulas.}
  \item{\dots}{potentially further arguments passed to methods.}
}
\description{
  This is method for formulas of the \code{\link[base]{merge}} generic
  function.  Here it is support for the function \code{\link{l1ce}} and
  not intended to be called directly by users.
}
\keyword{utilities}
% Converted by Sd2Rd version 1.21.
