% Copyright (C) 1998
% Berwin A. Turlach <bturlach@stats.adelaide.edu.au>
% Bill Venables <wvenable@stats.adelaide.edu.au>
% $Id: aux.l1celist.d,v 1.2 1999/12/01 05:08:22 bturlach Exp $
% --> ../COPYRIGHT for more details
\name{aux.l1celist}
\alias{aux.l1celist}
\title{Use `aux()' on a `l1celist' object}
\usage{
aux.l1celist(object, \dots)
}
\arguments{
  \item{object}{fitted model object (here typically of class
    \code{l1celist}).}
  \item{\dots}{potentilly further arguments passed to methods.}
}
\description{
  This is a method for the function \code{aux()} for objects inheriting
  from class \code{l1celist}.
  See \code{\link{aux}} for the general behavior of this function
  and for the interpretation of \code{object}.
}
\keyword{utilities}
% Converted by Sd2Rd version 1.21.
